requirements.txt
django>=4.2
requests>=2.31
yfinance>=0.2.40
pandas>=2.0
python-decouple>=3.8
